requirements.txt
captcha>=0.7.1
PyYAML>=6.0.2 